Refactored parts of storage.rs into temporary_channel.rs and creator_channel.rs.
Made storage more generic in case I want to use a different way of storing data later.

Found out what {} in main.rs does and changed it to just drop the mut var.

Started using the match keyword instead of if none return early.

Made a way to take a username (even if it is a special character) and get a synonym of room. (should go over later and check them as they were generated by the one and only LLM)

Added some tests that use special characters Ɽ.

Added permissions for created channels so that the creator can kick and move members, change the name, bitrate, max users in their channel