Improved on_voice_channel_join by returning an Option of a Result
Will now disconnect the user from the creator vc if something goes wrong

Temporary channels will now have the same bitrate and base permissions as their creator channels. With the permission addition of the temporary voice channel owner getting move members and manage channel permissions

Will now delete the temporary vc if not able to move the user to it.